refactor(testbed): layer the simulator's consumers above it
The consumer surface landed as loose modules in the source root, which
put the root both below the simulator (whose launcher, stub runtime, and
run internals import the root's Runtime contract, adapters, and
child-process helpers) and above it (the CLI, grader, and cc-judge
adapters importing simulator/*). That closed a folder cycle and took the
root to 20 children against a budget of 13.

The consumers move into two named boundaries with one direction of
knowledge: cli/ -> grading/ -> simulator/ -> the root contract. The
layer order is declared in the architecture config, so sibling-domain
crossings are governed by layering rather than waived one import at a
time.

grading/ holds grader.ts and the two cc-judge adapters together because
resolveRecordingHarness resolves the harness as a dist sibling of the
./grader entry; splitting them would break that resolution.

cli/main.ts rather than cli/index.ts: in this repo src/**/index.ts is a
curated public facade and carries the documentation rule set. A program
entry is not a facade.

The demo's spec fixture and its runner merge into cli/demo.ts. A canned
RunSpec is the demo verb's own business, and the indirection had one
consumer and two exports nothing imported.

grader.ts is declared a facade because the export map already names it
one; simulator/local-store.ts is declared a facade because
makeLocalRecordingStore now sits on the published ./simulator surface.
The root child budget moves 13 -> 15 for the two new named boundaries;
13 is unreachable without folding trace-capture-*, whose dist paths 16
eval scenarios name directly.

Also clears gates the consumer surface left red: dead exports knip
rejects, an orphaned documentStem, and one oxlint no-useless-spread.
