+++ packages/testbed/src/grading/cc-judge-recording-harness.ts
